refactor: user/enums/UserRole -> user/entity/UserRole
